added skipTag XmlPullParser util
Also moved safeNextString to XmlPullParser utils
